services/movie/requirements.txt
anyio==3.6.1
ariadne==0.15.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
graphql-core==3.2.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.20.4
typing-extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
